use embedded H2 database for testing

+++ src/main/java/com/example/message/repository/TodoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TodoRepository extends CrudRepository<Todo, Integer> {
+public interface TodoRepository extends CrudRepository<Todo, Long> {
 }
