+++ this-and-that/sql/postgres-setup.sql
@@ -7,7 +7,7 @@ create database tododb;
 
 grant all privileges on database tododb to todouser;
 
-create table todo (id SERIAL, description varchar(64) NOT NULL, priority INT, done BOOLEAN, primary key(id));
+create table todo (id BIGINT, description varchar(64) NOT NULL, priority INT, done BOOLEAN, primary key(id));
 
 grant all privileges on table todo to todouser;
